requirements.txt
oteapi-core>=0.1.3,<1
aiida-core>=2.0.1
DLite-Python>=0.3.9ver